Adding and browsing Ratings
Adding ratings in the user panel. Check if clicking on the star works, if you can only leave one rating as a buyer and seller of a given product.

Browsing ratings also for users who are not logged in. Enter e.g.:
http://127.0.0.1:8000/users/1
And see if the ratings match, if they are displayed correctly